+++ page-templates/homepage.php
@@ -11,5 +11,6 @@ get_header();
 get_template_part('template-parts/components/home-hero-slider');
 get_template_part('template-parts/components/interactive-select');
 get_template_part('template-parts/components/shows-cards-grid');
+get_template_part('template-parts/components/subscribe');
 
 get_footer();